Favorites
Added Favorites screen
Added multiplatform sql db
Updated movie id to Long

+++ androidApp/src/main/kotlin/dev/themobiledev/movie/navigation/MovieRoute.kt
@@ -3,5 +3,5 @@ package dev.themobiledev.movie.navigation
 sealed interface MovieRoute {
     data object PopularMoviesList : MovieRoute
     data object Favorites : MovieRoute
-    data class MovieDetail(val movieId: Int, val posterPath: String?, val title: String) : MovieRoute
+    data class MovieDetail(val movieId: Long, val posterPath: String?, val title: String) : MovieRoute
 }
