+++ sharedLogic/src/commonMain/kotlin/dev/themobiledev/movie/data/MoviesRepositoryImpl.kt
@@ -19,7 +19,7 @@ class MoviesRepositoryImpl(private val moviesApi: MoviesApi) : MoviesRepository 
     override suspend fun getNowPlayingMovies(page: Int): Result<MoviesPage> =
         moviesApi.getNowPlayingMovies(page).map { it.toDomain() }
 
-    override suspend fun getMovieDetails(movieId: Int): Result<MovieDetails> =
+    override suspend fun getMovieDetails(movieId: Long): Result<MovieDetails> =
         moviesApi.getMovieDetails(movieId).map { it.toDomain() }
 
     override suspend fun searchMovies(query: String, page: Int): Result<MoviesPage> =
