+++ sharedLogic/src/commonMain/kotlin/dev/themobiledev/movie/domain/MoviesRepository.kt
@@ -5,6 +5,6 @@ interface MoviesRepository {
     suspend fun getUpcomingMovies(page: Int = 1): Result<MoviesPage>
     suspend fun getTopRatedMovies(page: Int = 1): Result<MoviesPage>
     suspend fun getNowPlayingMovies(page: Int = 1): Result<MoviesPage>
-    suspend fun getMovieDetails(movieId: Int): Result<MovieDetails>
+    suspend fun getMovieDetails(movieId: Long): Result<MovieDetails>
     suspend fun searchMovies(query: String, page: Int = 1): Result<MoviesPage>
 }
